packaged auth in a blueprint
