+++ app/__init__.py
@@ -12,7 +12,7 @@ basedir = os.path.abspath(os.path.dirname(__file__))
 bootstrap = Bootstrap()
 moment = Moment()
 login = LoginManager()
-login.login_view = 'login'
+login.login_view = 'auth.login'
 
 db = SQLAlchemy()
 
@@ -29,6 +29,9 @@ def create_app(config_name):
     
     from .main import main as main_blueprint
     app.register_blueprint(main_blueprint)
+    
+    from .auth import auth as auth_blueprint
+    app.register_blueprint(auth_blueprint, url_prefix='/auth')
     
     return app
