+++ app/templates/base.html
@@ -18,10 +18,10 @@
             <ul class="nav navbar-nav">
                 <li><a href="/">Home</a></li>
                 {% if current_user.is_anonymous %}
-                <li><a href="/register">Register</a></li>
-                <li><a href="/login">Login</a></li>
+                <li><a href="{{url_for('auth.register')}}">Register</a></li>
+                <li><a href="{{url_for('auth.login')}}">Login</a></li>
                 {% else %}
-                <li><a href="/logout">Logout</a></li>
+                <li><a href="{{url_for('auth.logout')}}">Logout</a></li>
                 {% endif %}
                 
             </ul>
